Pop the stack on chained implicits and accumulate errors
When we fail to find a chained implicit in `typeclassTree`
we still want to pop the stack, otherwise the logic is broken.
On the other hand we want to keep the failed stack frames
in order to preserve the context for building a nice error message.
The problem is that `c.inferImplicitValue` is silent.

We solve the problem by keeping error frames in a separate stack.
When building the stack trace we put failed frames back on top.

Also factor out `magnolia:` prefix of error messages.