Add scaladoc targeted at newcomers to type-derivation
As a fairly experienced programmer, but a newcomer to type-derivation
and the terminology around Algebraic Data Types, I felt uncertain quite
quickly when trying to use Magnolia, and would have found some Scaladoc
helpful!

I'm not assuming that a developer would have to access rendered HTML
scaladoc to get value from this additional documentation - for instance,
the way I normally consume scaladoc is clicking through to the original
library source code in my IDE, and reading the scaladoc next to the
code.

In particular, I found these things puzzling as a newcomer, and have
tried to provide explanation/context (which may be wrong!) on them:

* What's the full list of methods needed to implement a derivation?
* Why are they called join & split?
* What would a typical way to implement them be - where are the examples of usage?

Some of the descriptions in the Scaladoc may look a little obvious or
unnecessary to experienced users of the library - the wording may appear
almost equivalent to the type signature of the method, but I do think
spelling these things verbally is helpful to new users!